Everywhere: Separate common helpers from InterpolationUtils
Helper datastructures representing partitions and interpolation colors
are separate entities and should live in separate files.
Interpolation tree is not currently used anywhere.
