+++ src/CMakeLists.txt
@@ -15,7 +15,6 @@ add_subdirectory(smtsolvers)
 add_subdirectory(parsers)
 add_subdirectory(itehandler)
 add_subdirectory(rewriters)
-add_subdirectory(interpolation)
 add_subdirectory(unsatcores)
 if (BUILD_EXECUTABLES)
   add_subdirectory(bin)
