--- src/interpolation/CMakeLists.txt
@@ -1,1 +1,0 @@
-install(FILES InterpolationUtils.h DESTINATION ${INSTALL_HEADERS_DIR}/interpolation)
